requirements.txt
redis==5.0.7
pydantic==2.8.2
pyyaml==6.0.2
structlog==24.1.0
prometheus-client==0.20.0
onnxruntime==1.18.1
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
fakeredis==2.23.2
numpy>=1.24
pandas>=2.2
pyarrow>=14
lightgbm>=4.3
scikit-learn>=1.4
onnxmltools>=1.11
onnxconverter-common>=1.14
matplotlib>=3.8
